refactor: extract interface for `DiskUsageMonitor`
Allows us to swap out the disk usage monitor with alternative
implementations.
